refactor: ResultRow 분리 및 memo 적용
